Plot State graphs. Prune on continuous activity.
The Viterbi algorithm was utilized to generate state graphs. The workers'
disagreement rate was plotted, while if they were in the good state or not
was plotted in the background. Also, the data is now also pruned based on
whether the activity is continuous or not. The results are saved to a csv